+++ examples/CustomerGetExample.java
@@ -8,9 +8,9 @@ public final class CustomerGetExample {
     public static void main(String[] args) throws Exception {
         String customerId = ExampleHelper.requiredEnv("CUSTOMER");
 
-        try (JammClient client = ExampleHelper.createClientFromEnv()) {
+        ExampleHelper.run((JammClient client) -> {
             Customer response = client.customers().get(customerId);
             ExampleHelper.printProto(response);
-        }
+        });
     }
 }