+++ examples/PaymentOnSessionAddChargeExample.java
@@ -23,9 +23,9 @@ public final class PaymentOnSessionAddChargeExample {
                 .build())
             .build();
 
-        try (JammClient client = ExampleHelper.createClientFromEnv()) {
+        ExampleHelper.run((JammClient client) -> {
             OnSessionPaymentResponse response = client.payments().onSessionPayment(request);
             ExampleHelper.printProto(response);
-        }
+        });
     }
 }