+++ examples/PaymentOnSessionContractWithChargeExample.java
@@ -27,9 +27,9 @@ public final class PaymentOnSessionContractWithChargeExample {
                 .build())
             .build();
 
-        try (JammClient client = ExampleHelper.createClientFromEnv()) {
+        ExampleHelper.run((JammClient client) -> {
             OnSessionPaymentResponse response = client.payments().onSessionPayment(request);
             ExampleHelper.printProto(response);
-        }
+        });
     }
 }