+++ examples/PaymentOnSessionContractWithoutChargeExample.java
@@ -17,9 +17,9 @@ public final class PaymentOnSessionContractWithoutChargeExample {
                 .build())
             .build();
 
-        try (JammClient client = ExampleHelper.createClientFromEnv()) {
+        ExampleHelper.run((JammClient client) -> {
             OnSessionPaymentResponse response = client.payments().onSessionPayment(request);
             ExampleHelper.printProto(response);
-        }
+        });
     }
 }