+++ examples/PaymentOnSessionOneTimePaymentExample.java
@@ -28,9 +28,9 @@ public final class PaymentOnSessionOneTimePaymentExample {
                 .build())
             .build();
 
-        try (JammClient client = ExampleHelper.createClientFromEnv()) {
+        ExampleHelper.run((JammClient client) -> {
             OnSessionPaymentResponse response = client.payments().onSessionPayment(request);
             ExampleHelper.printProto(response);
-        }
+        });
     }
 }